Исправляет подключение жс как модуль, разбивает фаил мэйн жс на модули дата жс и утил жс.
